+++ tools/xcpclient/src/main.rs
@@ -716,11 +716,26 @@ async fn xcp_client(
                     if segment_relative { "segment relative" } else { "absolute" }
                 );
 
-                // Register segments and event creations found in the code
-                elf_reader.register_segments_and_events(&mut reg, segment_relative, verbose)?;
-                // Find event triggers in the code and register their location (compilation unit, function, CFA offset)
+                // Get the EPK string and address from debug_data and set it in the registry application version information, if available
+                if elf_reader.debug_data.epk_addr > 0 {
+                    info!("EPK segment memory section found at address = 0x{:08X}", elf_reader.debug_data.epk_addr);
+                    let epk = elf_reader.debug_data.epk_string.clone().unwrap_or_else(|| "<unknown>".to_string());
+                    info!("EPK string: '{}'", epk);
+                    reg.application.set_version(epk, elf_reader.debug_data.epk_addr.try_into().unwrap());
+                } else {
+                    warn!("EPK segment memory section not found in ELF file");
+                }
+
+                // Register events from event creation markers (evt__name) in the code
+                elf_reader.register_events(&mut reg, verbose)?;
+
+                // Register segments from segment creation markers (calseg__name) found in the code
+                elf_reader.register_segments(&mut reg, segment_relative, verbose)?;
+
+                // Find event triggers in the code by their trigger markers (trg__mode__name) and register their location (compilation unit, function, CFA offset)
                 elf_reader.register_event_locations(&mut reg, verbose)?;
-                // Register accessible variables and their types
+
+                // Register all accessible variables and their types
                 elf_reader.register_variables(&mut reg, segment_relative, verbose, elf_idx_unit_limit)?; // register only variables <= compilation unit 0
             }